Refactored layers to required "opting-in" to each layer via a define.
Some keymaps will want different layers.  The problem with re-using the
same layer enum across multiple keymaps it forces keymap entries for all
layers, regardless if they are used or not.  This wastes space and
poses some risk if one of these unused layers is ever accidentally
activates.

Instead, guard each layer+keycode with a macro and require each keymap
to opt-in.  This requires a bit more boiler plate but is better than the
wasted space and undefined layer risk.